video-suggestions/requirements.txt
python-dotenv
pymongo 
scikit-learn 
flask